+++ Graphics/Src/AnimationIO.cpp
@@ -14,7 +14,7 @@ void AnimationIO::SaveAnimationClip(FILE * file, const AnimationClip & animClip)
 	for (uint32_t i = 0; i < animClip.boneAnimations.size(); ++i)
 	{
 		fprintf(file, "BoneName: %s\n", animClip.boneAnimations[i]->boneName.c_str());
-		fprintf(file, "PosKeyframes: %u\n", animClip.boneAnimations[i]->animation.mPosKeyframes.size());
+		fprintf(file, "PosKeyframes: %u\n", (int)animClip.boneAnimations[i]->animation.mPosKeyframes.size());
 		for (uint32_t index = 0; index < animClip.boneAnimations[i]->animation.mPosKeyframes.size(); ++index)
 		{
 			fprintf(file, "%f %f %f %f\n", animClip.boneAnimations[i]->animation.mPosKeyframes[index].mTime
@@ -22,7 +22,7 @@ void AnimationIO::SaveAnimationClip(FILE * file, const AnimationClip & animClip)
 				, animClip.boneAnimations[i]->animation.mPosKeyframes[index].mValue.y
 				, animClip.boneAnimations[i]->animation.mPosKeyframes[index].mValue.z);
 		}
-		fprintf(file, "QuatKeyframes: %u\n", animClip.boneAnimations[i]->animation.mQuatKeyframes.size());
+		fprintf(file, "QuatKeyframes: %u\n", (int)animClip.boneAnimations[i]->animation.mQuatKeyframes.size());
 		for (uint32_t index = 0; index < animClip.boneAnimations[i]->animation.mQuatKeyframes.size(); ++index)
 		{
 			fprintf(file, "%f %f %f %f %f\n", animClip.boneAnimations[i]->animation.mQuatKeyframes[index].mTime
@@ -31,7 +31,7 @@ void AnimationIO::SaveAnimationClip(FILE * file, const AnimationClip & animClip)
 				, animClip.boneAnimations[i]->animation.mQuatKeyframes[index].mValue.z
 				, animClip.boneAnimations[i]->animation.mQuatKeyframes[index].mValue.w);
 		}
-		fprintf(file, "ScaleKeyframes: %u\n", animClip.boneAnimations[i]->animation.mScaleKeyframes.size());
+		fprintf(file, "ScaleKeyframes: %u\n", (int)animClip.boneAnimations[i]->animation.mScaleKeyframes.size());
 		for (uint32_t index = 0; index < animClip.boneAnimations[i]->animation.mScaleKeyframes.size(); ++index)
 		{
 			fprintf(file, "%f %f %f %f\n", animClip.boneAnimations[i]->animation.mScaleKeyframes[index].mTime
@@ -49,7 +49,7 @@ void AnimationIO::LoadAnimationClip(FILE * file, std::vector<AnimationClip> & an
 	while (true)
 	{
 		char buffer[256];
-		int res = fscanf_s(file, "%s", buffer, std::size(buffer));
+		int res = fscanf_s(file, "%s", buffer, (int)std::size(buffer));
 
 		if (res == EOF)
 			break;
@@ -61,7 +61,7 @@ void AnimationIO::LoadAnimationClip(FILE * file, std::vector<AnimationClip> & an
 		}
 		else if (strcmp(buffer, "AnimationName:") == 0)
 		{
-			fscanf_s(file, "%s\n", buffer, std::size(buffer));
+			fscanf_s(file, "%s\n", buffer, (int)std::size(buffer));
 			animClip[++animIndex].name = std::string(buffer);
 			animClip[animIndex].boneAnimations.resize(numBones);
 			boneIndex = 0u;
@@ -76,13 +76,13 @@ void AnimationIO::LoadAnimationClip(FILE * file, std::vector<AnimationClip> & an
 		}
 		else if (strcmp(buffer, "BoneName:") == 0)
 		{
-			fscanf_s(file, "%s\n", buffer, std::size(buffer));
+			fscanf_s(file, "%s\n", buffer, (int)std::size(buffer));
 			animClip[animIndex].boneAnimations[boneIndex] = std::make_unique<BoneAnimation>();
 			animClip[animIndex].boneAnimations[boneIndex]->boneName += buffer;
 			AnimationBuilder builder;
 			for (int i = 0; i < 3; ++i)
 			{
-				fscanf_s(file, "%s", buffer, std::size(buffer));
+				fscanf_s(file, "%s", buffer, (int)std::size(buffer));
 				if (strcmp(buffer, "PosKeyframes:") == 0)
 				{
 					uint32_t num;
@@ -91,7 +91,7 @@ void AnimationIO::LoadAnimationClip(FILE * file, std::vector<AnimationClip> & an
 					{
 						Math::Vector3 v;
 						float t;
-						fgets(buffer, std::size(buffer), file);
+						fgets(buffer, (int)std::size(buffer), file);
 						if (sscanf_s(buffer, "%f %f %f %f",
 							&t, &v.x, &v.y, &v.z) != 4)
 						{
@@ -111,7 +111,7 @@ void AnimationIO::LoadAnimationClip(FILE * file, std::vector<AnimationClip> & an
 					{
 						Math::Quaternion q;
 						float t;
-						fgets(buffer, std::size(buffer), file);
+						fgets(buffer, (int)std::size(buffer), file);
 						if (sscanf_s(buffer, "%f %f %f %f %f", &t, &q.x, &q.y, &q.z, &q.w) != 5)
 						{
 							ASSERT(false, "[AnimationIO] Failed to load Rotation in Model file.");
@@ -130,7 +130,7 @@ void AnimationIO::LoadAnimationClip(FILE * file, std::vector<AnimationClip> & an
 					{
 						Math::Vector3 v;
 						float t;
-						fgets(buffer, std::size(buffer), file);
+						fgets(buffer, (int)std::size(buffer), file);
 						if (sscanf_s(buffer, "%f %f %f %f", &t, &v.x, &v.y, &v.z) != 4)
 						{
 							ASSERT(false, "[AnimationIO] Failed to load Scale in Model file.");
